Refactor and add Pagination component to shared UI library
