+++ packages/ui/src/Pagination/index.ts
@@ -1,0 +1,4 @@
+import { Pagination } from './Pagination'
+import { usePagination } from './usePagination'
+
+export { Pagination, usePagination }
