+++ packages/ui/src/Pagination/utils/types.ts
@@ -1,0 +1,2 @@
+export type PageQuery = string | undefined
+export type AnyObject = Record<string, unknown>
